Update Package JSON & resolve issue #6 (#9)
* init: Initial Commit

This is the initialisation of the project, I made some headway on completion of Issue #3. I am not yet ready to mark this issue as complete as it is just simply displaying a blank index.html (https://i.imgur.com/fB762lC.png) page.

* chore: Update package JSON files & gitignore

This commit updates the package.json file so that it uses electron-builder to compile my project from JavaScript to an executable file.
It also updates the .gitignore file. Resolves #6