feat: Adding support for multiple HTML files & small fixes.
This application now has support for multiple pages. This will be useful within the application to segment different sections and make the application easy to navigate.
There was also a new CONST variable added to the index.ts file, this should be toggled for deployment. This may be switched to using process.env in the future to make integration with CI/CD more effective. This resolves #8