Resolve Git Desynchronization & Resolves #12 (#13)
* init: Initial Commit

This is the initialisation of the project, I made some headway on completion of Issue #3. I am not yet ready to mark this issue as complete as it is just simply displaying a blank index.html (https://i.imgur.com/fB762lC.png) page.

* chore: Update package JSON files & gitignore

This commit updates the package.json file so that it uses electron-builder to compile my project from JavaScript to an executable file.
It also updates the .gitignore file. Resolves #6

* Update Package JSON & resolve issue #6 (#9)

* init: Initial Commit

This is the initialisation of the project, I made some headway on completion of Issue #3. I am not yet ready to mark this issue as complete as it is just simply displaying a blank index.html (https://i.imgur.com/fB762lC.png) page.

* chore: Update package JSON files & gitignore

This commit updates the package.json file so that it uses electron-builder to compile my project from JavaScript to an executable file.
It also updates the .gitignore file. Resolves #6

* feat: Adding support for multiple HTML files & small fixes.

This application now has support for multiple pages. This will be useful within the application to segment different sections and make the application easy to navigate.
There was also a new CONST variable added to the index.ts file, this should be toggled for deployment. This may be switched to using process.env in the future to make integration with CI/CD more effective. This resolves #8

* feat: Making further work on the HTML & implementing bootstrap. Resolves #3

* feat: Handle the building of the application automatically Resolves #12

* chore: remove debug console outputs

---------

Co-authored-by: Liam Townsley <contact@liamtownsley.me>